+++ src/main/java/com/dandelion/backend/payloads/dto/ProductDTO.java
@@ -18,7 +18,7 @@ public class ProductDTO {
 
     private String name;
 
-    private BigDecimal weight;
+    private Double weight;
 
     private Integer quantity;
